+++ backend/src/player.go
@@ -9,8 +9,8 @@ import (
 type Player struct {
 	game       *Game
 	id         int
-	xPos       int
-	yPos       int
+	xPos       float64
+	yPos       float64
 	angle      int
 	eventQueue []*PlayerEvent
 	alive      bool
@@ -43,8 +43,8 @@ func (p *Player) move(moveSpeed float64, moveAngle int) {
 	}
 
 	fmt.Printf("Player moving at speed %f at angle %d\n", moveSpeed, moveAngle)
-	p.xPos = int(moveSpeed*globalMoveSpeed*math.Cos(float64(moveAngle)*math.Pi/180.0)) + p.xPos
-	p.yPos = int(moveSpeed*globalMoveSpeed*math.Sin(float64(moveAngle)*math.Pi/180.0)) + p.yPos
+	p.xPos = moveSpeed*globalMoveSpeed*math.Cos(float64(moveAngle)*math.Pi/180.0) + p.xPos
+	p.yPos = moveSpeed*globalMoveSpeed*math.Sin(float64(moveAngle)*math.Pi/180.0) + p.yPos
 	p.angle = moveAngle
 
 }
@@ -55,7 +55,7 @@ func (p *Player) shoot(shotAngle int) {
 	}
 
 	fmt.Printf("Player shooting at angle %d\n", shotAngle)
-	currShot := &Shot{p.game.shotsFired + 1, p, int(float64(p.xPos) + math.Cos(float64(shotAngle)*math.Pi/180.0)*globalShotSpeed), int(float64(p.yPos) + math.Sin(float64(shotAngle)*math.Pi/180.0)*globalShotSpeed), shotAngle}
+	currShot := &Shot{p.game.shotsFired + 1, p, p.xPos + math.Cos(float64(shotAngle)*math.Pi/180.0)*globalShotSpeed, p.yPos + math.Sin(float64(shotAngle)*math.Pi/180.0)*globalShotSpeed, shotAngle}
 	p.game.shots = append(p.game.shots, currShot)
 	p.game.shotsFired++
